Moves zoom functions into separate scripts
