+++ module.json
@@ -11,6 +11,7 @@
       "name": "Syrious",
       "discord": "syrious_",
       "flags": {
+        "patreon": "SyriousWorkshop",
         "github": "Syrious",
         "ko-fi": "syrious"
       }
